server_a/requirements.txt
amqp==5.0.6
asgiref==3.4.1
AxmlParserPY==0.0.4
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
Brotli==1.0.9
cachetools==4.2.2
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.2
cent==4.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
cssselect2==0.4.1
cssutils==2.3.0
cycler==0.10.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.5
django-appconf==1.0.4
django-ckeditor==6.1.0
django-compressor==2.4.1
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-hosts==4.0
django-hvad==1.8.0
django-js-asset==1.2.2
django-parler==2.2
django-parler-rest==2.1
django-passwords==0.3.12
django-redis==5.0.0
django-smtp-ssl==1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Fabric3==1.14.post1
factory-boy==3.2.0
Faker==8.9.1
filelock==3.3.0
fonttools==4.31.2
hashids==1.3.1
hiredis==2.0.0
html5lib==1.1
idna==2.7
importlib-metadata==6.8.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
kiwisolver==1.3.1
kombu==5.1.0
lxml==4.6.3
Markdown==3.5.1
MarkupSafe==2.0.0
matplotlib==3.4.2
mock==4.0.3
mongoengine==0.23.1
nodeenv==1.6.0
numpy==1.21.0
packaging==21.0
paramiko==2.7.2
pbr==5.6.0
pdfrw==0.4
pika==1.3.2
Pillow==8.3.1
platformdirs==2.4.0
premailer==3.9.0
prompt-toolkit==3.0.19
psycopg2==2.9.3
psycopg2-binary==2.9.1
pyblake2==1.1.2
pycparser==2.20
pydyf==0.1.2
pymongo==3.11.4
PyNaCl==1.4.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2021.1
raven==6.10.0
rcssmin==1.0.6
redis==3.5.3
requests==2.20.1
rjsmin==1.1.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.2
sdh.ajax==1.0.11
sdh.forms==2.3.0
sdh.table==2.1.1
sentry-sdk==1.2.0
six==1.16.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
tinycss2==1.1.0
typing==3.7.4.3
ua2.otl==0.0.5
ua2.redis==0.0.4
uritemplate==3.0.1
urllib3==1.24.3
uWSGI==2.0.19.1
vine==5.0.0
virtualenv==20.8.1
wcwidth==0.2.5
weasyprint==54.0
webencodings==0.5.1
XlsxWriter==1.4.4
zipp==3.17.0
zopfli==0.2.1
